Almost done with word replacement!
Only problem I have now is that when I encounter the first object, it deletes one char of the current sentence. It's because  the remove function is supposed to start at word 5 to remove, but since it's not that big the first time, we get errors. Another thing, I need to find a way to store the old words with the word itself gone, unless we are supposed to keep it active somehow, mayeb just move the objects itself down to the sentence, then it's not really a problem. Other than that, it's working.