Tried improving
Added more items. Tried implementing swapping, where if there already is one item on the slot, it will get replaced. Think I'm very close to solving it, problem now is that it register it's occupied immideitly. Meaning that it's trying to move itself down back to the inventory, although it just got JUST moved there into an empty spot. Cuz now it's occupied. Something that can solve this is to have that it doesn't move or register the item before after you have released the mouse button. Although, it likely will see that its occupied by itself, and try and move itself back. Need some sort of test that can detect if it is its own item or someone else trying to get in there.