a lot of adding and deleting and failing on the script but i have an idea now
instead of having it sendt to one shroom then back again to the other shroom.
One can have like. The exit point furter away  and then a entry point. So EXIT ENTRY and then the other shroom ENTRY EXIT. So it will never hit where the other one exits. So that it will only be sendt to on place